requirements.txt
# Core dependencies
langchain>=0.0.181
pydantic>=1.10.8
fastapi>=0.100.0
uvicorn>=0.20.0
pytest>=8.0.0
pytest-asyncio>=0.25.0

# Audio processing and transcription
whisper>=1.1.10
deepgram-sdk>=3.10.0
pydub>=0.25.1

# NLP and Text Processing
spacy>=3.7.0
nltk>=3.9.0

# Vector storage
chromadb>=0.6.0

# Graph storage
neo4j>=5.22.0
py2neo>=2021.2.3

# LLM integrations
openai>=1.0.0
anthropic>=0.10.0

# Utilities
tqdm>=4.65.0
loguru>=0.7.0
pyyaml>=6.0.0